+++ src/main/java/skillapi/impl/client/gui/DefaultIcon.java
@@ -14,25 +14,25 @@ public class DefaultIcon implements ISkillIcon
 	}
 
 	@Override
-	public int u()
+	public double u()
 	{
 		return 0;
 	}
 
 	@Override
-	public int U()
+	public double U()
 	{
 		return 1;
 	}
 
 	@Override
-	public int v()
+	public double v()
 	{
 		return 0;
 	}
 
 	@Override
-	public int V()
+	public double V()
 	{
 		return 1;
 	}
